+++ src/index/mod.rs
@@ -5,7 +5,6 @@ mod cell;
 mod edge;
 mod iterator;
 mod mode;
-mod triangle;
 mod vertex;
 
 pub use cell::CellIndex;
@@ -14,4 +13,3 @@ pub use mode::IndexMode;
 pub use vertex::{Vertex, VertexIndex};
 
 pub use iterator::{Children, GridPathCells};
-use triangle::Triangle;